Refactor test names and remove flaky test (#403)
* Refactor test names and remove flaky test

* Fix naming

* Rename tests (again)
