+++ march_safety/test/connection_not_lost_test.cpp
@@ -6,28 +6,35 @@
 #include <std_msgs/Time.h>
 
 #include <march_safety/temperature_safety.h>
-#include "ErrorCounter.cpp"
+#include "error_counter.h"
 
-class TestConnectionNeverStarted : public ::testing::Test
+class TestConnectionNotLost : public ::testing::Test
 {
 protected:
 };
 
-TEST_F(TestConnectionNeverStarted, connectionNeverStarted)
+TEST_F(TestConnectionNotLost, connectionNotLost)
 {
   ros::Time::init();
   ros::NodeHandle nh;
+  int send_errors_interval;
+  nh.getParam("/march_safety_node/send_errors_interval", send_errors_interval);
+  ros::Publisher pub_alive = nh.advertise<std_msgs::Time>("march/input_device/alive", 0);
   ErrorCounter errorCounter;
   ros::Subscriber sub = nh.subscribe("march/error", 0, &ErrorCounter::cb, &errorCounter);
 
-  while (0 == sub.getNumPublishers())
+  while (0 == pub_alive.getNumSubscribers() || 0 == sub.getNumPublishers())
   {
     ros::Duration(0.1).sleep();
   }
+  EXPECT_EQ(1, pub_alive.getNumSubscribers());
   EXPECT_EQ(1, sub.getNumPublishers());
 
+  std_msgs::Time timeMessage;
+  timeMessage.data = ros::Time::now();
+  pub_alive.publish(timeMessage);
   ros::spinOnce();
-  ros::Duration(0.5).sleep();
+  ros::Duration(send_errors_interval / 2 / 1000).sleep();
   ros::spinOnce();
   EXPECT_EQ(0, errorCounter.count);
 }
