+++ march_safety/test/error_counter.h
@@ -1,4 +1,5 @@
 // Copyright 2019 Project March.
+#include <march_shared_resources/Error.h>
 
 struct ErrorCounter
 {
@@ -6,10 +7,10 @@ struct ErrorCounter
   {
   }
 
-  void cb(const march_shared_resources::Error& msg)
+  void cb(const march_shared_resources::Error&)
   {
     ++count;
   }
 
-  uint32_t count;
+  size_t count;
 };
