+++ march_safety/test/march_safety_connection_never_started.test
@@ -10,5 +10,5 @@
         </rosparam>
     </node>
 
-    <test test-name="march_safety_test" pkg="march_safety" type="march_safety-test-connection-never-started"/>
+    <test test-name="march_safety_test" pkg="march_safety" type="march_safety_connection_never_started_test"/>
 </launch>
