+++ march_safety/test/march_safety_connection_not_lost.test
@@ -10,5 +10,5 @@
         </rosparam>
     </node>
 
-    <test test-name="march_safety_test" pkg="march_safety" type="march_safety-test-connection-not-lost"/>
+    <test test-name="march_safety_test" pkg="march_safety" type="march_safety_connection_not_lost_test"/>
 </launch>
