+++ march_safety/test/march_safety_temperature.test
@@ -21,5 +21,5 @@
         </rosparam>
     </node>
 
-    <test test-name="march_safety_test" pkg="march_safety" type="march_safety-test"/>
+    <test test-name="march_safety_test" pkg="march_safety" type="march_safety_temperature_test"/>
 </launch>
